Revert "Include labels in the output of /inspect."
This reverts commit 4efe057ecb800ac8e7aaf9c75aa6654aa5535c7f.

By adding extra "naming" data to each probe, I was making it impossible
for the trace router to find the data it wanted, so transformations all
broke. I may find a better way to get these labels in the output of
/inspect eventually, but at the moment it looks pretty hopeless. Since
it's not a critical feature, @lancepantz may just have to do without.