Make it easier to specify storage retentions.
Added 'time-span function to relieve repetition, and an 'every macro
to make the time-spans read like english text. Usually I'm not a fan of
trying to make code look like natural language, but here it's a small
amount of code, and the english does a good job of providing context
when reading that would be missing if it looked more like a function call.