Improve subscription-filter feature.
Instead of this weird plan of encoding information into the name key,
and using subscription filters to muck with the name, we now have
separate name and target settings, implemented in telemetry/server
instead of in each module. So you can have target be *, meaning
"write this to a file corresponding exactly to the group-by key",
but still have a unique name to distinguish it from other queries that
want to write to a corresponding filename.